requirements.txt
#backend_service @ git+https://github.com/Sheesh1006/proto.git@6926fb502090a5fc76224358c6069f72519307be#subdirectory=gen/backend_service
git+https://@github.com/Sheesh1006/proto.git@main#subdirectory=gen/backend_service
blinker==1.9.0
click==8.1.8
Flask==3.1.0
grpcio==1.71.0
grpcio-tools==1.71.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
munch==4.0.0
protobuf==5.29.4
PyYAML==6.0.2
setuptools==78.1.0
Werkzeug==3.1.3
